requirements.txt
asgiref==3.4.1
beautifulsoup4==4.11.1
cobble==0.1.3
Django==3.2.7
django-bootstrap-v5==1.0.11
django-ckeditor==6.1.0
django-crispy-forms==1.12.0
django-js-asset==1.2.2
django-taggit==1.5.1
mammoth==1.4.19
Pillow==9.1.1
pytz==2021.1
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
Unidecode==1.3.1
xlrd==1.2.0
